// types/profile.ts
export type Profile = {
  name: string;
  title: string;
  avatar?: unknown;
  shortBio?: string;
  longBio?: unknown[];
  location?: string;
  email?: string;
  githubUrl?: string;
  linkedinUrl?: string;
  facebookUrl?: string;
  resumeFile?: unknown;
  resumeUrl?: string;
  heroCtaPrimary?: string;
  heroCtaSecondary?: string;
};